connections page update
