+++ admin/settings/class-connections.php
@@ -9,25 +9,6 @@ class LNP_ConnectionPage extends LNP_SettingsPage
     protected $template_html = 'settings/page-connections.php';
     protected $option_name   = 'lnp_connection';
 
-    public function init_fields()
-    {
-        parent::init_fields();
-
-        $this->lnd_section();
-
-        $this->lndbits_section();
-
-        $this->btcpay_section();
-
-        $this->lnaddress_section();
-
-        $this->lndhub_section();
-
-        $this->lndhub_create_account_section();
-
-        $this->navigation();
-    }
-
     /**
      * Make menu item/page title translatable
      */
@@ -39,321 +20,223 @@ class LNP_ConnectionPage extends LNP_SettingsPage
     }
 
 
-    public function navigation()
+    /**
+     * Register Tabs if any
+     * @return [type] [description]
+     */
+    public function init_fields()
     {
-        $section = 'connection-types';
-        $this->add_section([
-            'key' => $section,
-            'title' => ''
-        ]);
-
-        $this->add_custom_field([
-            'key' => 'connection_type_lnd',
-            'name' => null,
-            'section' => $section
-        ], array($this, 'field_connection_types'));
-    }
-
-    public function lnd_section()
-    {
-        $section = 'lnd';
-
-        $this->add_section([
-            'key' => $section,
-            'title' => 'LND Config'
-        ]);
-
-        $this->add_input_field([
-            'key' => 'lnd_address',
-            'name' => 'Address',
-            'label' => 'e.g. https://127.0.0.1:8080 - or <a href="#" id="load_from_lndconnect">click here to load details from a lndconnect</a>',
-            'section' => $section,
-            'type' => 'url'
-        ]);
-
-        $this->add_input_field([
-            'key' => 'lnd_macaroon',
-            'name' => 'Macaroon',
-            'label' => 'Invoices macaroon in HEX format',
-            'section' => $section
-        ]);
-
-        $this->add_input_field([
-            'key' => 'lnd_cert',
-            'name' => 'TLS Cert',
-            'label' => 'TLS Certificate',
-            'section' => $section
-        ]);
-    }
-
-    public function lndbits_section()
-    {
-        $section = 'lnbits';
-
-        $this->add_section([
-            'key' => $section,
-            'title' => 'LNbits Config'
-        ]);
-
-
-        $this->add_input_field([
-            'key' => 'lnbits_apikey',
-            'name' => 'API Key',
-            'label' => 'LNbits Invoice/read key',
-            'section' => $section,
-        ]);
-
-        $this->add_input_field([
-            'key' => 'lnbits_host',
-            'name' => 'Host',
-            'label' => 'LNbits host (e.g. https://legend.lnbits.com)',
-            'section' => $section,
-            'type' => 'url'
-        ]);
-    }
-
-    public function btcpay_section()
-    {
-        $section = 'btcpay';
-
-        $this->add_section([
-            'key' => $section,
-            'title' => 'BTC Payer'
-        ]);
-
-        $this->add_input_field([
-            'key' => 'btcpay_host',
-            'name' => 'Host',
-            'label' => 'BtcPay Host',
-            'section' =>  $section,
-            'type' => 'url'
-        ]);
-
-        $this->add_input_field([
-            'key' => 'btcpay_apikey',
-            'name' => 'ApiKey',
-            'label' => 'BtcPay Api Key',
-            'section' => $section
-        ]);
-
-        $this->add_input_field([
-            'key' => 'btcpay_store_id',
-            'name' => 'Store Id',
-            'label' => 'BtcPay Store Id',
-            'section' => $section
-        ]);
-    }
-
-    public function lnaddress_section()
-    {
-        $section = 'lnaddress';
-
-        $this->add_section([
-            'key' => $section,
-            'title' => 'Lightning Address Config'
-        ]);
-
-        $this->add_input_field([
-            'key' => 'lnaddress_address',
-            'name' => 'Lightning Address',
-            'label' => 'Lightning Address (e.g. you@payaddress.co) - only works if the vistor supports WebLN!',
-            'section' => $section,
-        ]);
-    }
-
-    public function lndhub_section()
-    {
-        $section = 'lndhub';
-
-        $this->add_section([
-            'key' => $section,
-            'title' => 'LndHub Config'
-        ]);
-
-        $this->add_input_field([
-            'key' => 'lndhub_url',
-            'name' => 'LndHub Url',
-            'label' => 'LndHub Host',
-            'section' => $section,
-            'type' => 'url'
-        ]);
-
-        $this->add_input_field([
-            'key' => 'lndhub_login',
-            'name' => 'LndHub Login',
-            'label' => '',
-            'section' => $section
-        ]);
-
-        $this->add_input_field([
-            'key' => 'lndhub_password',
-            'name' => 'LndHub Password',
-            'label' => '',
-            'section' => $section,
-            'type' => 'password'
-        ]);
-    }
-
-    public function lndhub_create_account_section()
-    {
-        $section = 'lndhub_create_account';
-
-        $this->add_section([
-            'key' => $section,
-            'title' => 'LndHub Config'
-        ]);
-
-        $this->add_input_field([
-            'key' => 'lndhub_url',
-            'name' => 'LndHub Url',
-            'label' => 'LndHub Host',
-            'section' => $section,
-            'type' => 'hidden',
-        ]);
-
-        $this->add_input_field([
-            'key' => 'lndhub_login',
-            'name' =>   'LndHub Login',
-            'type' => 'hidden',
-            'label' => '',
-            'section' => $section
-        ]);
-
-        $this->add_input_field([
-            'key' => 'lndhub_password',
-            'name' => 'LndHub Password',
-            'label' => '',
-            'section' => $section,
-            'type' => 'hidden',
-        ]);
-
-        $this->add_custom_input_field([
-            'key' => 'lndhub_create_button',
-            'name' => 'LndHub Create Wallet',
-            'label' => 'Create Wallet',
-            'section' => $section,
-        ], array($this, 'field_create_account_section'));
-    }
-
-    public function field_create_account_section($args)
-    {
-        printf(
-            '<div>
-                <button id="%s" type="button" class="button button-danger">Create wallet</button>
-             </div>',
-            'create_lndhub_account',
-            $args['label']
+        
+        // Tabs
+        $this->tabs = array(
+            'lnd' => array(
+                'title'       => __('LNDhub', 'lnp-alby' ),
+                'description' => __('Connect using LNDhub', 'lnp-alby'),
+            ),
+            'lndhub' => array(
+                'title'       => __('LndHub (BlueWallet)', 'lnp-alby' ),
+                'description' => __('Connect using LNDhub', 'lnp-alby'),
+            ),
+            'lnbits' => array(
+                'title'       => __('LNbits', 'lnp-alby' ),
+                'description' => __('Connect to your LNbits account', 'lnp-alby'),
+            ),
+            'btcpay' => array(
+                'title'       => __('BTC Pay', 'lnp-alby' ),
+                'description' => __('Connect using BTCPay Server', 'lnp-alby'),
+            ),
+            'lnaddress' => array(
+                'title'       => __('LN Address', 'lnp-alby' ),
+                'description' => __('Connect using Lightning Address Config', 'lnp-alby'),
+            ),
         );
-    }
 
-    public function field_connection_types()
-    {
-        echo '<div class="wp-lnp-card__container">';
-        $this->connection_card('lnd', $this->plugin->get_file_url('assets/img/lnd.png'), 'Lnd', 'Connect using lndhub');
-        $this->connection_card('lndhub', $this->plugin->get_file_url('assets/img/lndhub.png'), 'LndHub (BlueWallet)', 'Connect using lndhub');
-        $this->connection_card('lnbits', $this->plugin->get_file_url('assets/img/lnbits.png'), 'LNbits', 'Connect to your LNbits account');
-        $this->connection_card('lnaddress', $this->plugin->get_file_url('assets/img/satsymbol-black.png'), 'Lightning Address Config', 'Connect using Lightning Address Config');
-        $this->connection_card('btcpay', $this->plugin->get_file_url('assets/img/BTCPay_Icon_with_background.png'), 'BTC Pay', 'Connect using BTCPay Server');
-        $this->connection_card('lndhub_create_account', $this->plugin->get_file_url('assets/img/alby.png'), 'Create a new wallet', 'We create and manage a lightning wallet for you');
-        echo '</div>';
+        parent::init_fields();
     }
 
 
-    public function connection_card($id, $image, $title, $subtitle)
-    {
-        printf(
-            '<div data-section="%s" class="wp-lnp-card">
-                <div class="wp-lnp-card__header">
-                    <img src="%s" class="wp-lnp-card__image" />
-                </div>
-                <div class="wp-lnp-card__body">
-                    <h4 class="wp-lnp-card__title">%s</h4>
-                    <p class="wp-lnp-card__text">%s</p>
-                </div>
-            </div>',
-            $id,
-            $image,
-            $title,
-            $subtitle
+    /**
+     * Array of form fields available on this page
+     */
+    public function set_form_fields() {
+
+        /**
+         * Fields 
+         */
+        $fields = array();
+
+        /**
+         * Fields for section: LND Hub
+         */
+        $fields[] = array(
+            'tab'     => 'lnd',
+            'field'   => array(
+                'label'       => __( 'Address', 'lnp-alby' ),
+                'name'        => 'lnd_address',
+                'type'        => 'url',
+                'description' => __( 'e.g. https://127.0.0.1:8080 - or <a href="#" id="load_from_lndconnect">click here to load details from a lndconnect</a>', 'lnp-alby'),
+            ),
         );
+
+        $fields[] = array(
+            'tab'     => 'lnd',
+            'field'   => array(
+                'label'       => __( 'Macaroon', 'lnp-alby' ),
+                'name'        => 'lnd_macaroon',
+                'description' => __( 'Invoices macaroon in HEX format', 'lnp-alby'),
+            ),
+        );
+
+        $fields[] = array(
+            'tab'     => 'lnd',
+            'field'   => array(
+                'label'       => __( 'TLS Certificate', 'lnp-alby' ),
+                'name'        => 'lnd_cert',
+                'description' => __( 'TLS Certificate', 'lnp-alby'),
+            ),
+        );
+
+
+        /**
+         * Fields for section: LND Hub (Blue wallet)
+         */
+        $fields[] = array(
+            'tab'     => 'lndhub',
+            'field'   => array(
+                'label'       => __( 'LndHub Url', 'lnp-alby' ),
+                'name'        => 'lndhub_url',
+                'type'        => 'url',
+                'description' => __( 'LndHub Host', 'lnp-alby'),
+            ),
+        );
+
+        $fields[] = array(
+            'tab'     => 'lndhub',
+            'field'   => array(
+                'label'       => __( 'Login', 'lnp-alby' ),
+                'name'        => 'lndhub_login',
+                'description' => __( 'LndHub Login', 'lnp-alby'),
+            ),
+        );
+
+        $fields[] = array(
+            'tab'     => 'lndhub',
+            'field'   => array(
+                'label'       => __( 'Password', 'lnp-alby' ),
+                'name'        => 'lndhub_password',
+                'type'        => 'password',
+                'description' => __( 'LndHub Password', 'lnp-alby'),
+            ),
+        );
+
+        /**
+         * LNBits
+         */
+        $fields[] = array(
+            'tab'     => 'lnbits',
+            'field'   => array(
+                'label'       => __( 'API Key', 'lnp-alby' ),
+                'name'        => 'lnbits_apikey',
+                'description' => __( 'LNbits Invoice/read key', 'lnp-alby'),
+            ),
+        );
+
+        $fields[] = array(
+            'tab'     => 'lnbits',
+            'field'   => array(
+                'label'       => __( 'Host', 'lnp-alby' ),
+                'name'        => 'lnbits_host',
+                'type'        => 'url',
+                'description' => __( 'LNbits host (e.g. https://legend.lnbits.com)', 'lnp-alby'),
+            ),
+        );
+
+
+        /**
+         * BTC Pay
+         */
+        $fields[] = array(
+            'tab'     => 'btcpay',
+            'field'   => array(
+                'label'       => __( 'Host', 'lnp-alby' ),
+                'name'        => 'btcpay_host',
+                'type'        => 'url',
+                'description' => __( 'BtcPay Host', 'lnp-alby'),
+            ),
+        );
+
+        $fields[] = array(
+            'tab'     => 'btcpay',
+            'field'   => array(
+                'label'       => __( 'API Key', 'lnp-alby' ),
+                'name'        => 'btcpay_apikey',
+                'description' => __( 'BtcPay Api Key', 'lnp-alby'),
+            ),
+        );
+
+        $fields[] = array(
+            'tab'     => 'btcpay',
+            'field'   => array(
+                'label'       => __( 'Store ID', 'lnp-alby' ),
+                'name'        => 'btcpay_store_id',
+                'description' => __( 'BtcPay Store ID', 'lnp-alby'),
+            ),
+        );
+
+
+        /**
+         * LN Address
+         */
+        $fields[] = array(
+            'tab'     => 'lnaddress',
+            'field'   => array(
+                'label'       => __( 'Lightning Address', 'lnp-alby' ),
+                'name'        => 'lnaddress_address',
+                'description' => __( 'Lightning Address (e.g. you@payaddress.co) - only works if the vistor supports WebLN!', 'lnp-alby'),
+            ),
+        );
+
+        // Save Form fields to class
+        $this->form_fields = $fields;
     }
 
-    public function render_current_section($page)
-    {
-        global $wp_settings_sections;
 
-        if (!isset($wp_settings_sections[$page])) {
-            return;
+
+    public function get_lnd_address_node_info() {
+
+        try {
+
+            if (
+                $this->plugin->getLightningClient()
+                && $this->plugin->getLightningClient()->isConnectionValid()
+            ) {
+                
+                $node_info = $this->plugin->getLightningClient()->getInfo();
+                
+                $type    = 'notice';
+                $message = sprintf(
+                    '%s %s - %s',
+                    __('Connected to:', 'lnp-alby'),
+                    $node_info['alias'],
+                    $node_info['identity_pubkey']
+                );
+            }
+            else {
+
+                $type    = 'error';
+                $message = __('Not connected', 'lnp-alby');
+            }
+        }
+        catch (Exception $e) {
+
+            $type    = 'error';
+            $message = sprintf(
+                '%s %s',
+                __('Not connected', 'lnp-alby'),
+                $e
+            );
         }
 
-        $type = $this->plugin->lightningClientType;
-        $i = 0;
-        foreach ((array) $wp_settings_sections[$page] as $section) {
-
-            if (!isset($type) && $i !== 0) continue;
-
-            if (isset($type) && $type !== $section['id']) continue;
-
-            $this->render_section($page, $section, false);
-            break;
-        }
-    }
-
-    function do_settings_sections($page)
-    {
-        global $wp_settings_sections, $wp_settings_fields;
-
-        if (!isset($wp_settings_sections[$page])) {
-            return;
-        }
-
-
-        foreach ((array) $wp_settings_sections[$page] as $section) {
-            if ($section['id'] === 'connection-types') continue;
-            $this->render_section($page, $section, true);
-        }
-    }
-
-    public function render_section($page, $section, $hide)
-    {
-        global $wp_settings_fields;
-        $class = $hide ? 'wp-lnp-section__hidden' : '';
-
-        echo "<div id='{$section['id']}' class='wp-lnp-section {$class}'>";
-
-        if ($section['title']) {
-            echo "<h2>{$section['title']}</h2>\n";
-        }
-
-        if ($section['callback']) {
-            call_user_func($section['callback'], $section);
-        }
-
-        if (!isset($wp_settings_fields) || !isset($wp_settings_fields[$page]) || !isset($wp_settings_fields[$page][$section['id']])) {
-            return;
-        }
-
-        if ($section['id'] === 'connection-types' || $section['id'] === 'lndhub_create_account') {
-            echo '<div>';
-            $this->do_settings_fields($page, $section['id']);
-            echo '</div>';
-        } else {
-            echo '<table class="form-table" role="presentation">';
-            do_settings_fields($page, $section['id']);
-            echo '</table>';
-        }
-        echo '</div>';
-    }
-
-    public function do_settings_fields($page, $section)
-    {
-        global $wp_settings_fields;
-
-        if (!isset($wp_settings_fields[$page][$section])) {
-            return;
-        }
-        foreach ((array) $wp_settings_fields[$page][$section] as $field) {
-            call_user_func($field['callback'], $field['args']);
-        }
+        $this->add_admin_notice($message, $type);
     }
 }
